% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dputtruncat.R
\name{dput_grist}
\alias{dput_grist}
\title{Object to be dput-ed}
\usage{
dput_grist(n = 25, left = 4, right = 8)
}
\arguments{
\item{n}{a number of numbers}

\item{left}{smallest number}

\item{right}{largest number}
}
\value{
character vector
}
\description{
An random assortment of characters.
}
\examples{
dput_grist()
}
